backend/requirement.txt
# --- Web Framework & Server ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2

# --- MQTT & VDA5050 ---
paho-mqtt>=1.6.1
git+https://github.com/zeki-aitech/vda5050_client.git

# --- Security & Database ---
bcrypt>=4.0.1
cryptography>=41.0.0
PyJWT>=2.8.0

# --- Utilities ---
pyyaml>=6.0